+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add CODE_OF_CONDUCT.md
+- Add StorageHealthCheck
 
 ### Changed