+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.0] - 2020-10-03
+
 ### Added
 
 - Add CODE_OF_CONDUCT.md
@@ -108,7 +110,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial commit
 
-[unreleased]: https://github.com/ukfast/laravel-health-check/compare/v1.4.0...HEAD
+[unreleased]: https://github.com/ukfast/laravel-health-check/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/ukfast/laravel-health-check/tree/v1.5.0
 [1.4.0]: https://github.com/ukfast/laravel-health-check/tree/v1.4.0
 [1.3.0]: https://github.com/ukfast/laravel-health-check/tree/v1.3.0
 [1.2.1]: https://github.com/ukfast/laravel-health-check/tree/v1.2.1